Update README.md
I wasn't able to see the logs in Grafana because user promtail didn't have the permissions to read my caddy logs. So I added a P.S. as a reminder.
Deleted an additional tab that wasn't supposed to be there.
Added the yaml tag to some examples for better readability.